add test for cone activation
This is still not good. At the moment, this actually implements a hyper-pyramid, not a cone.
It would be better to create a cone information (slope, offset), then test the expansion factor
not individually but on the multi-d test data.
Cones (components) should only be active when the area is much smaller than a rectangle.