avoid long integrations
previously, the next likelihood threshold marked for updating the
strategy (checking whether the integration is done)
was placed at the second-to-last live point.

In some problems, when frac_remain is low (e.g., 50%), this lead to
unnecessarily cautious and expensive continuation of the integration.

This updates that the marker is placed to the median live
point likelihood, at most.