simplify and correct rank order testing
Simulations of uniform random ranks indicate
that runs shorter than 300000 do not occur at
a 4 sigma threshold (at 3 sigma, ~10% are arbitrarily short).

This implies that with 100000 NS iterations, no more than 1 runs,
and with 1000000 iterations no more than 3 runs are expected.

However, when small biases are introduced,
such as sampling only 395 of 400, or with a rank powerlaw probability
with power index 0.02 (0 is uniform), 100% of the time
more than the expected runs are created an thus detectable
with 100000 iterations. With 1000000, 0.01 and 396/400 is always
detectable.

This commit implements this as the test, by
- altering the threshold from 2 to 4
- only counting the main iterator runs (because the bootstrapped
  ones are highly correlated and thus not useful independent
  indicators)
- implementing the nruns < niter / 10^(5.5) significance rule