Loop over `run` folders to find next available
Rather than adding up the number of objects in the log directory to work out the next one, this makes the code loop from 1 to find the first available directory of the form `runX`. There's a a hard maximum of 1000, to guarantee termination, after which the code raises an error. Feel free to change this (or make it a parameter): I'm not sure if anyone might want to store more than 1000 runs.

You may also want to raise a different kind of error.

This is intended to resolve issue #33