implement better logz error improvement rule. speed up tests.
Previously, the logz range to improve was obtained by sampling.
But the error is due to all points before the bulk,
so a minimum number of live points needs to be introduced.

The new strategy identifies a minimum number of live points to add
to achieve the desired error. In the case of a constant nlive run,
this is sigma(lnZ) = sqrt(iterations / nlive**2),
as the variance in each compression is gamma with 1/nlive mean and uncertainty.

In changing nlive this is not as trivial, so the
variance sum terms 1/nlive**2 need to be identified.
The total variance is bounded by the largest terms, i.e., the smallest
nlive. So the code iterates nlive to see what minimum nlive
would be needed to achieve the target.

Running all tests is getting a bit slow, and this commit identifies
the slowest tests, which can be skipped with pytest -k 'SLOW'