keep minimum number of live points set by number of clusters for longer
Previously, when nlive < cluster_num_live_points * nclusters,
the iteration was stopped and the required number of live
points was set from parents of the current live points
until the current Lmin.
On the next iteration, once Lmin was reached, the number
of live points decreased in each iteration.
At some point, the criterion is hit again, and the process repeats.
This creates many iterations, and a slow-down, because each
iteration integrates from the start (although no new live point
evaluations).

This change sets the required number of live points until Lmax,
the current highest likelihood.
This avoids doing too many iterations.

A potential drawback is that Lmax is actually very high, ie.,
well past the point where the clusters cease to exist.