when running out of constraint, revert to earliest known good point
MPI runs with a stepsampler can violate detailed balance and be biased.

The stepsampler can end up with some of its trajectory
out of the likelihood constraint, when it is raised.

Reverting only to the last good point retains non-reversible
trajectories.

This patch reverts to the first good point (if any), instead,
ensuring reversibility.

This patch also removes the AHARM step sampler, which was never
completed and is superseded by another branch.