skip live point test if using step sampler
When using a step sampler with few live points in a high-d scenario,
it can occur that the ellipsoid moves and does not cover the live points.
This is because the ellipsoid is continuously updated with the
live point center, but may be too small (especially in d>10).
However, when a step sampler is used which ignores the region
(region filtering off), this is inconsequential.

This commit turns the check off.

For MLFriends runs, the check remains, but I have not yet seen any
problem there.

It can be turned off by running python with optimization (-O etc),
which disables all asserts.