first attempt at projecting slice sampler
Here, the live points are projected onto the proposed slice
to guess the slice length.
I tried various ways of guessing.
However, already the least optimized one leads to biases.
For that one, I take the interval of projected live points, and
extend it by its length in both directions.
Padding even more (by a constant, 8) gives correct result,
but is not more effient. Test case was a Gaussian in 4d and 16d.