switch default to new local covariance, LocalAffineLayer
LocalAffineLayer uses the MLFriends distance to identify neighbours,
and subtracts the neighbour mean from each point before computing
a covariance. In a L-shape for example, the neighbour covariance
computed in the I part and the _ part can come out different,
and would be averaged. This is different to AffineLayer,
which takes the "diagonal", which is not optimal.
L-shape likelihood contours occur with additive multi-component fitting,
where the normalisations have log-uniform priors.

Based on this, LocalAffineLayer should lead to a covariance that
better captures local covariance in a way that is useful to MLFriends.

Indeed, in the rosenbrock example, a slightly better sampling
efficiency can be observed with the LocalAffineLayer than AffineLayer.

However, for a Gaussian likelihood, LocalAffineLayer is
sampling slightly less efficiently than AffineLayer.
This can be understood because AffineLayer uses all points to build
the covariance, and thereby probably gets a estimate of the covariance
that converges better to the truth.
The sampling efficiency however is already high for ellipsoidal contours
such as from Gaussians, partially supported by the encircling ellipsoid
in the MLFriends class.

This commit therefore probably helps over a wider class of problems,
while compromising with slightly poorer performance for ellipsoidal
contours.