remove the feature of reverting the step sampler.
Old behaviour:
When running with MPI, step samplers can progress at different efficiencies.
The point of the first one that finishes N steps is accepted and this increases
the nested sampling likelihood threshold. The other step samplers are informed
and may revert / back track to an earlier state. This can cause some areas of
the parameter space to be explored much slower, and thus become undersampled, leading to biases.
see https://github.com/JohannesBuchner/UltraNest/issues/159

New behaviour:
Reverting is disabled. The step sampler remembers the Lmin it was started
with, and returns the Lmin->L edge to the ReactiveNestedSampling integrator,
which stores it and may use it (unless the new Lmin is already above L).

This should reduce the issue, but is still not perfect, as the edge may
arrive too late. Ideally, we should wait until all stepsamplers at the
current Lmin have finished, but this is difficult to synchronize.

Resuming could pick up the late edges; however, we would need to sort
the point store by Lmin. This should be technically possible within
an improvement loop if we keep track of the last pointstore entry
at the start of the loop, and keep the entries since then sorted.