undoing a previous commit, allowing reverting stepsamplers again
Without reverting, a stepsampler may return a result, but it cannot
be used anymore, even when resuming it would probably not be loaded.
With reverting, raising the Lmin should affect only the stepsamper
with probability 1/K, so the sampler has a good chance of not having
to revert back very often.

With the new queue, variation in executation duration should be less
of an issue.